Debug and Fix on the ICIFAR Dataloader
Major updates:
Transform is now applied only before the subset so we can return
prototypes without augmentation
Splitted Get X of class in two methods, one to get the images
to be used as prototypes (no augm. or normalization)
and one to get a DataLoader to compute the method response.

TODO:
- IDAD still needs to be updated to respect the new flow.
- Main must be refactored and must use config files or different main scripts
- ICaRL must be updated to keep into account the N + M + M flow.